+++ MoL/time_integrators.f90
@@ -1,5 +1,4 @@
 module time_integrators
-  use iso_fortran_env, only: real64
   use geometry,        only: geometry_t
   use mesh,            only: mesh_t
   use state,           only: state_t
@@ -11,7 +10,7 @@ module time_integrators
   public :: bc_init_from_geometry, bc_set_rgrid
 
   ! ---- Almacenamiento opcional de coeficientes de BC / malla ----
-  real(real64), allocatable, save :: rgrid(:), alpha(:), bta(:), guu(:)
+  real(kind=8), allocatable, save :: rgrid(:), alpha(:), bta(:), guu(:)
 
 contains
 
@@ -26,7 +25,7 @@ contains
   end subroutine bc_init_from_geometry
 
   subroutine bc_set_rgrid(r)
-    real(real64), intent(in) :: r(:)
+    real(kind=8), intent(in) :: r(:)
     call assign_or_copy(rgrid, r)
   end subroutine bc_set_rgrid
 
@@ -35,18 +34,17 @@ contains
   !============================
   subroutine ssprk3_step_state(dt, S, G, M, ell, mu, &
                               phi_p, psi_p, pi_p)
-    use iso_fortran_env, only: real64
     use rhs,             only: compute_rhs_state
     implicit none
-    real(real64),  intent(in)    :: dt
+    real(kind=8),  intent(in)    :: dt
     type(state_t), intent(inout) :: S
     type(geometry_t), intent(in) :: G
     type(mesh_t),     intent(in) :: M
     integer,          intent(in) :: ell
-    real(real64),     intent(in) :: mu
+    real(kind=8),     intent(in) :: mu
 
-    complex(real64), intent(inout) :: phi_p(:), psi_p(:), pi_p(:)
-    complex(real64) :: dphi(S%Nr), dpsi(S%Nr), dpi(S%Nr)
+    complex(kind=8), intent(inout) :: phi_p(:), psi_p(:), pi_p(:)
+    complex(kind=8) :: dphi(S%Nr), dpsi(S%Nr), dpi(S%Nr)
 
     integer :: N, rk
 
@@ -68,20 +66,20 @@ contains
         S%pi  = pi_p  + dt*dpi
 
       case (2)
-        S%phi = (3.0_real64/4.0_real64)*phi_p + &
-                (1.0_real64/4.0_real64)*(S%phi + dt*dphi)
-        S%psi = (3.0_real64/4.0_real64)*psi_p + &
-                (1.0_real64/4.0_real64)*(S%psi + dt*dpsi)
-        S%pi  = (3.0_real64/4.0_real64)*pi_p  + &
-                (1.0_real64/4.0_real64)*(S%pi  + dt*dpi)
+        S%phi = (3.0d0/4.0d0)*phi_p + &
+                (1.0d0/4.0d0)*(S%phi + dt*dphi)
+        S%psi = (3.0d0/4.0d0)*psi_p + &
+                (1.0d0/4.0d0)*(S%psi + dt*dpsi)
+        S%pi  = (3.0d0/4.0d0)*pi_p  + &
+                (1.0d0/4.0d0)*(S%pi  + dt*dpi)
 
       case (3)
-        S%phi = (1.0_real64/3.0_real64)*phi_p + &
-                (2.0_real64/3.0_real64)*(S%phi + dt*dphi)
-        S%psi = (1.0_real64/3.0_real64)*psi_p + &
-                (2.0_real64/3.0_real64)*(S%psi + dt*dpsi)
-        S%pi  = (1.0_real64/3.0_real64)*pi_p  + &
-                (2.0_real64/3.0_real64)*(S%pi  + dt*dpi)
+        S%phi = (1.0d0/3.0d0)*phi_p + &
+                (2.0d0/3.0d0)*(S%phi + dt*dphi)
+        S%psi = (1.0d0/3.0d0)*psi_p + &
+                (2.0d0/3.0d0)*(S%psi + dt*dpsi)
+        S%pi  = (1.0d0/3.0d0)*pi_p  + &
+                (2.0d0/3.0d0)*(S%pi  + dt*dpi)
       end select
 
       !call bc_apply_state(S) 
@@ -92,15 +90,14 @@ contains
 
 
 subroutine bc_apply_state(S, G, M)
-  use iso_fortran_env, only: real64
   implicit none
   class(state_t),    intent(inout) :: S
   class(geometry_t), intent(in)    :: G
   class(mesh_t),     intent(in)    :: M
 
   integer :: N
-  real(real64) :: sN
-  complex(real64) :: wplus_bnd, wminus_bnd
+  real(kind=8) :: sN
+  complex(kind=8) :: wplus_bnd, wminus_bnd
 
   N = size(S%phi)
   if (N < 4) stop "bc_apply_state: N < 4 no soportado"
@@ -108,12 +105,12 @@ subroutine bc_apply_state(S, G, M)
   !===========================================================
   ! 1) Frontera interna (i = 1): extrapolación 2º orden
   !===========================================================
-  !S%phi(1) = S%phi(2) !3.0_real64*S%phi(2) - 3.0_real64*S%phi(3) + S%phi(4)
-  !S%psi(1) = S%psi(2) !3.0_real64*S%psi(2) - 3.0_real64*S%psi(3) + S%psi(4)
-  !S%pi(1)  = S%pi(2)  !3.0_real64*S%pi(2)  - 3.0_real64*S%pi(3)  + S%pi(4)
-S%phi(1) = 2.0_real64*S%phi(2) - S%phi(3)
-S%psi(1) = 2.0_real64*S%psi(2) - S%psi(3)
-S%pi(1)  = 2.0_real64*S%pi(2)  - S%pi(3)
+  !S%phi(1) = S%phi(2) !3.0d0*S%phi(2) - 3.0d0*S%phi(3) + S%phi(4)
+  !S%psi(1) = S%psi(2) !3.0d0*S%psi(2) - 3.0d0*S%psi(3) + S%psi(4)
+  !S%pi(1)  = S%pi(2)  !3.0d0*S%pi(2)  - 3.0d0*S%pi(3)  + S%pi(4)
+S%phi(1) = 2.0d0*S%phi(2) - S%phi(3)
+S%psi(1) = 2.0d0*S%psi(2) - S%psi(3)
+S%pi(1)  = 2.0d0*S%pi(2)  - S%pi(3)
   S%phi(N) = S%phi(N-1)
   S%psi(N) = S%psi(N-1)
   S%pi(N)  = S%pi(N-1)
@@ -128,8 +125,8 @@ end subroutine bc_apply_state
   ! Utilidad para copiar/asignar arrays reales allocatables
   !============================
   subroutine assign_or_copy(a, b)
-    real(real64), allocatable, intent(inout) :: a(:)
-    real(real64),           intent(in)       :: b(:)
+    real(kind=8), allocatable, intent(inout) :: a(:)
+    real(kind=8),           intent(in)       :: b(:)
     if (allocated(a)) then
       if (size(a) /= size(b)) then
         deallocate(a)
